+++ lib/guess_os/type.rb
@@ -29,11 +29,11 @@ module GuessOS
 
       output = conn.last_output
       items = output.split
-      type =  'gnu/linux'
+      type =  :gnulinux
       name =  items[1]&.downcase
       desc =  output
 
-      list = %w(debian ubuntu opensuse manjaro)
+      list = %w(debian ubuntu opensuse manjaro mint)
       unless list.include? name
         return OS.new(:unkown, :unkown, 'Unkown OS')
       end
@@ -45,8 +45,8 @@ module GuessOS
       conn = GuessOS::Conn.new(host)
       command = 'ver'
       conn.exec(command)
-      # command = 'echo %windir%' => Windows
       # ver => Microsoft Windows [Version 10.0.20348.469]
+      # command = 'echo %windir%' => Windows
 
       identified = conn.ok && conn.last_output.include?('Windows')
       return OS.new(:unkown, :unkown, conn.status) unless identified
@@ -56,7 +56,7 @@ module GuessOS
       output.gsub!("\n", '')
       items = output.split
 
-      type = 'windows'
+      type = :windows
       name = "windows #{items[3].split('.').first}"
       desc = output
       OS.new(type, name, desc)
@@ -71,9 +71,37 @@ module GuessOS
       return OS.new(:unkown, :unkown, conn.status) unless identified
 
       output = conn.last_output
-      type =  'cygwin'
+      type =  :cygwin
       name =  'cygwin'
       desc =  output.gsub("\n", '')
+      OS.new(type, name, desc)
+    end
+
+    def self.guess_macos(host)
+      conn = GuessOS::Conn.new(host)
+      command = 'ver'
+      conn.exec(command)
+      # 1.9.2 kevin-macbookpro:~ $ system_profiler SPSoftwareDataType
+      # System Version: Mac OS X 10.7.4 (11E53)
+      #
+      # 1.9.2 kevin-macbookpro:~ $ sw_vers
+      # ProductName:    Mac OS X
+      # ProductVersion: 10.7.4
+
+      command = 'sw_vers '
+      # ver => Microsoft Windows [Version 10.0.20348.469]
+
+      identified = conn.ok && conn.last_output.include?('Windows')
+      return OS.new(:unkown, :unkown, conn.status) unless identified
+
+      output = conn.last_output
+      output.gsub!("\r", '')
+      output.gsub!("\n", '')
+      items = output.split
+
+      type = :macos
+      name = "Mac OS #{items[3].split('.').first}"
+      desc = output
       OS.new(type, name, desc)
     end
 
@@ -88,7 +116,7 @@ module GuessOS
 
       output = conn.last_output
       items = output.split
-      type =  'minix'
+      type =  :minix
       name =  'minix'
       desc =  output.gsub("\n", '')
       OS.new(type, name, desc)