+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+## [0.1.1] - 2022-07-07
+
+- Change OS.type values from Strings to Symbols.
+- OS.type values = [ :gnulinux, :windows, :macos, :cygwin, :minix ]
+
 ## [0.1.0] - 2022-07-03
 
 - Initial release
+- Guess OS:
+